packages/plugin-crewai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neocortex-crewai"
version = "0.1.0"
description = "CrewAI integrations for TinyHumans memory API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "tinyhumansai>=0.5",
    "crewai>=0.28.0",
    "pydantic>=2.4.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
examples = ["crewai-tools", "python-dotenv>=1.0"]

[tool.setuptools.packages.find]
include = ["neocortex_crewai*"]

[tool.ruff]
line-length = 88
target-version = "py39"
